Refactored each class/function into a module, and same for their tests
