+++ metroscope/__init__.py
@@ -1,7 +1,7 @@
 """Initialize the metroscope package."""
 
-from .metroscope import WordBuilder
-from .metroscope import LineBuilder
-from .metroscope import scanned_poem
+from .WordBuilder import WordBuilder
+from .LineBuilder import LineBuilder
+from .scanned_poem import scanned_poem
 
 __all__ = ["WordBuilder", "LineBuilder", "scanned_poem"]
