+++ metroscope/test_WordBuilder.py
@@ -1,10 +1,6 @@
-"""
-Unit test the metroscope package.
+"""Unit test the WordBuilder Class."""
 
-Coverage complete, switching to Test Driven Development.
-"""
-
-from metroscope import WordBuilder, LineBuilder
+from metroscope import WordBuilder
 
 
 class Test_WordBuilder(object):
@@ -215,85 +211,3 @@ class Test_WordBuilder(object):
         for word, HTML in WORDS.items():
             METER = [0, 1, 0]
             assert(WordBuilder(word).stressed_HTML(METER) == HTML)
-
-
-class Test_LineBuilder(object):
-    """
-    Test the LineBuilder class.
-
-    Coverage is complete and this class is under TDD.
-    """
-
-    def test_init(self):
-        """Should initialize from the original line."""
-        LINES = (
-                 "One morn before me were three figures seen,",
-                 "And once more came they by:-alas! wherefore?",
-                 )
-        for line in LINES:
-            assert(LineBuilder(line).line == line)
-
-    def test_str_magic_method(self):
-        """Should return the original word."""
-        LINES = (
-                 "One morn before me were three figures seen,",
-                 "And once more came they by:-alas! wherefore?",
-                 )
-        for line in LINES:
-            assert(str(LineBuilder(line)) == line)
-
-    def test_repr_magic_method(self):
-        """Should evaluate to itself."""
-        LINES = (
-                 "One morn before me were three figures seen,",
-                 "And once more came they by:-alas! wherefore?",
-                 )
-        for line in LINES:
-            assert(repr(LineBuilder(line))
-                   == "LineBuilder('" + line + "')")
-
-    def test__clean_line(self):
-        """Should replace hyphens and emlines with spaces."""
-        LINES = {
-                 "One morn before me were three figures seen,":
-                 "One morn before me were three figures seen,",
-                 "And once—more came they by:-alas! wherefore?":
-                 "And once more came they by: alas! wherefore?",
-                 }
-        for line, clean_line in LINES.items():
-            assert(LineBuilder(line)._clean_line() == clean_line)
-
-    def test__word_list(self):
-        """Should create a list of WordBuilder instances."""
-        LINE = "Two Owls and a Hen,"
-        for word in LineBuilder(LINE)._word_list:
-            assert(repr(word) == "WordBuilder('" + str(word) + "')")
-
-    def test__rhyming_part(self):
-        """Should return the rhyming part of the last word of the line."""
-        LINES = {
-                 "Two Owls and a Hen,": "EH1 N",
-                 "I knew to be my demon Poesy.": None
-                 }
-        for line, rhyme in LINES.items():
-            assert(LineBuilder(line)._rhyming_part == rhyme)
-
-    def test_stressed_HTML(self):
-        """Should give an HTML representation of the line's fit to meter."""
-        LINES = {
-                 "Two Owls and a Hen,":
-                 "<span>\
-<span style='color:black'>Two</span></span> \
-<span><strong style='color:black'>Owls</strong></span> \
-<span><span style='color:black'>and</span></span> \
-<span><span style='color:black'>a</span></span> \
-<span><strong style='color:black'>Hen,</strong>\
-</span> ",
-                 }
-        for line, HTML in LINES.items():
-            METER = [0, 1, 0, 0, 1]
-            assert(LineBuilder(line).stressed_HTML(METER) == HTML)
-
-
-if __name__ == '__main__':
-    print("Run pytest from the command line.")
